+++ src/Libraries Projects/Milochau.Core.Tests/Infrastructure/Hosting/ApplicationHostEnvironmentTests.cs
@@ -12,11 +12,12 @@ namespace Milochau.Core.Tests.Infrastructure.Hosting
         private const string ApplicationName = "applicationName";
         private const string EnvironmentName = "Production";
         private const string HostName = "prd";
+        private const string RegionName = "ew1";
 
         [TestInitialize]
         public void Initialize()
         {
-            MilochauHostingEnvironment = new ApplicationHostEnvironment(OrganizationName, ApplicationName, EnvironmentName, HostName);
+            MilochauHostingEnvironment = new ApplicationHostEnvironment(OrganizationName, ApplicationName, EnvironmentName, HostName, RegionName);
         }
 
         [TestMethod("MilochauHostingEnvironment - Properties")]
@@ -26,6 +27,7 @@ namespace Milochau.Core.Tests.Infrastructure.Hosting
             Assert.AreEqual(ApplicationName, MilochauHostingEnvironment.ApplicationName);
             Assert.AreEqual(EnvironmentName, MilochauHostingEnvironment.EnvironmentName);
             Assert.AreEqual(HostName, MilochauHostingEnvironment.HostName);
+            Assert.AreEqual(RegionName, MilochauHostingEnvironment.RegionName);
         }
 
         [TestMethod("IsProduction - Called")]